change increment to bigincrement

+++ database/migrations/2016_12_28_111111_create_api_key_access_events_table.php
@@ -14,8 +14,8 @@ class CreateApiKeyAccessEventsTable extends Migration
     public function up()
     {
         Schema::create('api_key_access_events', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('api_key_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('api_key_id');
             $table->ipAddress('ip_address');
             $table->text('url');
             $table->timestamps();
